requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flask-wtf==1.2.1